Forge formatting

+++ test/mocks/TestSMRewardDistributor.sol
@@ -10,18 +10,11 @@ contract TestSMRewardDistributor is SMRewardDistributor {
         uint256 _maxRewardMultiplier,
         uint256 _smoothingValue,
         address _ecosystemReserve
-    )
-        SMRewardDistributor(
-            _safetyModule,
-            _maxRewardMultiplier,
-            _smoothingValue,
-            _ecosystemReserve
-        )
-    {}
+    ) SMRewardDistributor(_safetyModule, _maxRewardMultiplier, _smoothingValue, _ecosystemReserve) {}
 
     function accrueRewards(address user) external {
         uint256 numMarkets = _getNumMarkets();
-        for (uint i; i < numMarkets; ) {
+        for (uint256 i; i < numMarkets;) {
             _accrueRewards(_getMarketAddress(_getMarketIdx(i)), user);
             unchecked {
                 ++i;
